fix NESTML custom syntax highlighting
